trackgen: fix some gcc warnings
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@9460 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: fd70345df8c973d58684915e959fc2c3d39db234
Former-commit-id: e01c46defcd5f998639663f87bd9750caca25b74

+++ src/tools/trackgen/ac3d.cpp
@@ -1197,7 +1197,7 @@ void Ac3d::Object::generateNormals()
                 {
                     for (size_t j = 0; j < surface.refs.size(); j++)
                     {
-                        if (surface.refs[j].index == i)
+                        if (surface.refs[j].index == static_cast<int>(i))
                         {
                             if (j < 3)
                             {
